move cargo and emonhub_interfacer, clean emonhub.py part 1
